fix cert verification and update build file
The Certificate verification failed due to a type conversion and a
missing return statement. The npnt_check_authenticity method is modified
to correct for this.

npnt.h was missing the definition for the set_dgca_public_key
method. This has been fixed in this commit

build.gradle is updated to use pre-built raspberry pi.
WARN: the path is currently hardcoded. This needs to be changed to use
relative paths.